разбивание на компоненты, добавление маршрутизации
