+++ src/components/private-route/private-route.tsx
@@ -1,0 +1,19 @@
+import { Navigate } from 'react-router-dom';
+import { AppRoute, AuthorizationStatus } from '../../const';
+
+type PrivateRouteProps = {
+    authorizationStatus: AuthorizationStatus;
+    children: JSX.Element;
+}
+
+function PrivateRoute(props: PrivateRouteProps): JSX.Element{
+    const {authorizationStatus, children} = props;
+
+    return (
+        authorizationStatus === AuthorizationStatus.Auth
+        ? children
+        : <Navigate to={AppRoute.Login} />
+    )
+}
+
+export default PrivateRoute;
