+++ src/const.ts
@@ -1,0 +1,13 @@
+export enum AppRoute {
+    Main = '/',
+    Favorites = '/favorites',
+    Offer = '/offer/:id',
+    Login = '/login',
+    NotFound = '*'
+}
+
+export enum AuthorizationStatus {
+    Auth = 'AUTH',
+    NoAuth = 'NO_AUTH',
+    Unknown = 'UNKNOWN',
+}
