+++ src/pages/identical-section/card-offer/card-offer.tsx
@@ -1,3 +1,6 @@
+import { Link } from 'react-router-dom';
+import Bookmark from '../bookmark/bookmark';
+
 function CardOffer(): JSX.Element {
   return(
     <article className="cities__card place-card">
@@ -5,7 +8,7 @@ function CardOffer(): JSX.Element {
         <span>Premium</span>
       </div>
       <div className="cities__image-wrapper place-card__image-wrapper">
-        <a href="#">
+        <Link to="#">
           <img
             className="place-card__image"
             src="img/apartment-01.jpg"
@@ -13,7 +16,7 @@ function CardOffer(): JSX.Element {
             height={200}
             alt="Place image"
           />
-        </a>
+        </Link>
       </div>
       <div className="place-card__info">
         <div className="place-card__price-wrapper">
@@ -21,12 +24,7 @@ function CardOffer(): JSX.Element {
             <b className="place-card__price-value">€120</b>
             <span className="place-card__price-text">/&nbsp;night</span>
           </div>
-          <button className="place-card__bookmark-button button" type="button">
-            <svg className="place-card__bookmark-icon" width={18} height={19}>
-              <use xlinkHref="#icon-bookmark" />
-            </svg>
-            <span className="visually-hidden">To bookmarks</span>
-          </button>
+          <Bookmark />
         </div>
         <div className="place-card__rating rating">
           <div className="place-card__stars rating__stars">
