+++ src/pages/identical-section/cities-map/cities-map.tsx
@@ -1,0 +1,9 @@
+function CitiesMap(): JSX.Element {
+  return(
+    <div className="cities__right-section">
+      <section className="cities__map map"></section>
+    </div>
+  );
+}
+
+export default CitiesMap;
