+++ src/pages/identical-section/header-six-sities/header-six-sities.tsx
@@ -1,0 +1,11 @@
+function HeaderSixSities(): JSX.Element {
+  return(
+    <div className="header__left">
+      <a className="header__logo-link header__logo-link--active">
+        <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
+      </a>
+    </div>
+  );
+}
+
+export default HeaderSixSities;
