+++ src/pages/identical-section/premium/premium.tsx
@@ -1,0 +1,9 @@
+function Premium (): JSX.Element {
+  return(
+    <div className="place-card__mark">
+      <span>Premium</span>
+    </div>
+  );
+}
+
+export default Premium;
